backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
python-multipart
reportlab
pdfplumber
Pillow
google-generativeai
python-dotenv
